V2.1.3:  Update CI for Ubuntu 22.04 (#38)
GitHub Actions upgrading to Ubuntu 22 broke a few things because of
version differences/incompatibilities. This fixes them, and now
hardcodes all of the Ubuntu jobs to specify a specific version, .e.g,
`ubuntu-22.04` and never use `ubuntu-latest`, to prevent this in the
future. In the future we should manually upgrade things whenever a new
Ubuntu LTS is available.

- Fixes PR #24, which broke because a newer version of TBB removed a
deprecated function that we were using
- Uses a newer version of Google Benchmark which no longer spams error
messages during configuration