Elastic scheduler (#52)
This patch involves a substantial new feature for the work-stealing
scheduler. Now, if the scheduler fails to find enough work to saturate
the workers, they will go to sleep and wait until there is more work
available. This means that while the code is running only sequential
code, or is blocked on user input, the workers will not be consuming CPU
time.

This feature is enabled by default, since experiments show it to have no
substantial performance impact (bench_standard results are within 0.04%
on average). It can be disabled by setting the compile definition
PARLAY_ELASTIC_PARALLELISM to false. The time that workers will
attempt to steal before timing out and going to sleep can be controlled
with the macro PARLAY_ELASTIC_STEAL_TIMEOUT, which is in microseconds
(default is 10000, i.e., 10ms, and seems to work well).

This addresses the feature request #33 and subsumes PR #34 which solved
the same problem but resulted in a slowdown on the benchmarks. It also
contains the minor bug fix for #51.

---------

Co-authored-by: Guy Blelloch <guyb@cs.cmu.edu>